move stuff start on new lexer
