+++ tesyl/src/interpreter_old/parser_old.rs
@@ -1,4 +1,4 @@
-use crate::{ast::BinOp, Exp};
+use crate::{ast_old::BinOp, Exp};
 use core::panic;
 use std::iter::Peekable;
 use std::vec::IntoIter;
