+++ tesyl/src/position.rs
@@ -1,0 +1,4 @@
+pub struct Position {
+    line: u32,
+    char: u32,
+}
